% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.make.table.R
\name{internal.make.table}
\alias{internal.make.table}
\title{Experiments' Labels Predictor}
\usage{
internal.make.table(directory, columns, ExcludedSDBs)
}
\arguments{
\item{directory}{The location (path) of the experiment data.}

\item{columns}{The included columns of the experiment data, input as list. Ex. c(1,2)}

\item{ExcludedSDBs}{The excluded SDBs of the experiment data, input as list. Ex. c("SDB3", "SDB77")}
}
\value{
table
}
\description{
This package predict the label of the given experiment.
}
\examples{

}
