% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.predict.label.R
\name{internal.predict.label}
\alias{internal.predict.label}
\title{Experiments' Labels Predictor}
\usage{
internal.predict.label(test_dict, file_name)
}
\arguments{
\item{file_name}{}
}
\value{
table
}
\description{
This package predict the label of the given experiment.
}
\examples{

}
