% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.label.R
\name{predict.label}
\alias{predict.label}
\title{Experiments' Labels Predictor}
\usage{
\method{predict}{label}(table)
}
\arguments{
\item{table}{}
}
\value{
table
}
\description{
This package predict the label of the given experiment.
}
